Move swagger setup to own file
